requirements.txt
selenium==3.141.0
robotframework==4.0b2
